Fix NPE when updating unauthed room.
Weird: Static room data loads fine for unauthed rooms,
so we know what machines are there and everything. The
only thing that doesn't work is the request for dynamic
(cycle) data, which is being returned as an empty string.
Hence, unauthed rooms now load fine, but with every
machine grayed out.